update proxmox provider; update deploy to run locally and with cicd

+++ deploy.sh
@@ -7,18 +7,21 @@ ANSIBLE_VERSION=$(ansible --version | grep core | awk '{print $NF}' | sed 's/]//
 export PATH=/root/.pkgx/ansible.com/v${ANSIBLE_VERSION}/venv/bin:$PATH
 pip install netaddr
 
-echo "${ansible_pihole_vars_main_yml}" > ansible/pihole/vars/main.yml
-echo "${common_tfbackend}" > dns/dns-ha/config.s3.tfbackend
-echo "${common_tfbackend}" > dns/pihole/blue-config.s3.tfbackend
-echo "${common_tfbackend}" > dns/pihole/green-config.s3.tfbackend
-echo "${dnsdist_tfstate_key}" >> dns/dns-ha/config.s3.tfbackend
-echo "${blue_pihole_tfstate_key}" >> dns/pihole/blue-config.s3.tfbackend
-echo "${green_pihole_tfstate_key}" >> dns/pihole/green-config.s3.tfbackend
-echo "${dnsdist_tfvars}" > dns/dns-ha/terraform.tfvars
-echo "${pihole_blue_tfvars}" > dns/pihole/blue.tfvars
-echo "${pihole_green_tfvars}" > dns/pihole/green.tfvars
-echo "${pihole_default_tfvars}" > dns/pihole/terraform.tfvars
-echo "${dnsdist_ansible_vars}" > ansible/dnsdist/vars/main.yml
+
+if [ ! -f ".local" ]; then
+  echo "${ansible_pihole_vars_main_yml}" > ansible/pihole/vars/main.yml
+  echo "${common_tfbackend}" > dns/dns-ha/config.s3.tfbackend
+  echo "${common_tfbackend}" > dns/pihole/blue-config.s3.tfbackend
+  echo "${common_tfbackend}" > dns/pihole/green-config.s3.tfbackend
+  echo "${dnsdist_tfstate_key}" >> dns/dns-ha/config.s3.tfbackend
+  echo "${blue_pihole_tfstate_key}" >> dns/pihole/blue-config.s3.tfbackend
+  echo "${green_pihole_tfstate_key}" >> dns/pihole/green-config.s3.tfbackend
+  echo "${dnsdist_tfvars}" > dns/dns-ha/terraform.tfvars
+  echo "${pihole_blue_tfvars}" > dns/pihole/blue.tfvars
+  echo "${pihole_green_tfvars}" > dns/pihole/green.tfvars
+  echo "${pihole_default_tfvars}" > dns/pihole/terraform.tfvars
+  echo "${dnsdist_ansible_vars}" > ansible/dnsdist/vars/main.yml
+fi
 
 tofu --version
